Struct futures_timer::Delay

pub struct Delay { /* fields omitted */ }

A future representing the notification that an elapsed duration has occurred.

This is created through the Delay::new or Delay::new_at methods indicating when the future should fire at. Note that these futures are not intended for high resolution timers, but rather they will likely fire some granularity after the exact instant that they're otherwise indicated to fire at.

Methods

impl Delay

pub fn new(dur: Duration) -> Delay

Creates a new future which will fire at dur time into the future.

The returned object will be bound to the default timer for this thread. The default timer will be spun up in a helper thread on first use.

pub fn new_at(at: Instant) -> Delay

Creates a new future which will fire at the time specified by at.

The returned object will be bound to the default timer for this thread. The default timer will be spun up in a helper thread on first use.

pub fn new_handle(at: Instant, handle: TimerHandle) -> Delay

Creates a new future which will fire at the time specified by at.

The returned instance of Delay will be bound to the timer specified by the handle argument.

pub fn reset(&mut self, dur: Duration)

Resets this timeout to an new timeout which will fire at the time specified by dur.

This is equivalent to calling reset_at with Instant::now() + dur

pub fn reset_at(&mut self, at: Instant)

Resets this timeout to an new timeout which will fire at the time specified by at.

This method is usable even of this instance of Delay has "already fired". That is, if this future has resovled, calling this method means that the future will still re-resolve at the specified instant.

If at is in the past then this future will immediately be resolved (when poll is called).

Note that if any task is currently blocked on this future then that task will be dropped. It is required to call poll again after this method has been called to ensure tha ta task is blocked on this future.