Struct futures_timer::Interval

pub struct Interval { /* fields omitted */ }

A stream representing notifications at fixed interval

Intervals are created through the Interval::new or Interval::new_at methods indicating when a first notification should be triggered and when it will be repeated.

Note that intervals are not intended for high resolution timers, but rather they will likely fire some granularity after the exact instant that they're otherwise indicated to fire at.

Methods

impl Interval

pub fn new(dur: Duration) -> Interval

Creates a new interval which will fire at dur time into the future, and will repeat every dur interval after

The returned object will be bound to the default timer for this thread. The default timer will be spun up in a helper thread on first use.

pub fn new_at(at: Instant, dur: Duration) -> Interval

Creates a new interval which will fire at the time specified by at, and then will repeat every dur interval after

The returned object will be bound to the default timer for this thread. The default timer will be spun up in a helper thread on first use.

pub fn new_handle(at: Instant, dur: Duration, handle: TimerHandle) -> Interval

Creates a new interval which will fire at the time specified by at, and then will repeat every dur interval after

The returned object will be bound to the timer specified by handle.