Struct futures_timer::Timer

pub struct Timer { /* fields omitted */ }

A "timer heap" used to power separately owned instances of Delay and Interval.

This timer is implemented as a priority queued-based heap. Each Timer contains a few primary methods which which to drive it:

next_wake indicates how long the ambient system needs to sleep until it invokes further processing on a Timer

advance_to is what actually fires timers on the Timer, and should be called essentially every iteration of the event loop, or when the time specified by next_wake has elapsed.

The Future implementation for Timer is used to process incoming timer updates and requests. This is used to schedule new timeouts, update existing ones, or delete existing timeouts. The Future implementation will never resolve, but it'll schedule notifications of when to wake up and process more messages.

Note that if you're using this crate you probably don't need to use a Timer as there is a global one already available for you run on a helper thread. If this isn't desirable, though, then the TimerHandle::set_fallback method can be used instead!

Methods

impl Timer

pub fn new() -> Timer

Creates a new timer heap ready to create new timers.

pub fn handle(&self) -> TimerHandle

Returns a handle to this timer heap, used to create new timeouts.

pub fn next_event(&self) -> Option<Instant>

Returns the time at which this timer next needs to be invoked with advance_to.

Event loops or threads typically want to sleep until the specified instant.

pub fn advance(&mut self)

Proces any timers which are supposed to fire at or before the current instant.

This method is equivalent to self.advance_to(Instant::now()).

pub fn advance_to(&mut self, now: Instant)

Proces any timers which are supposed to fire before now specified.

This method should be called on Timer periodically to advance the internal state and process any pending timers which need to fire.