Crate futures_timer

A general purpose crate for working with timeouts and delays with futures.

Examples

use std::time::Duration;
use futures_timer::Delay;

let now = Delay::new(Duration::from_secs(3)).await;
println!("waited for 3 secs");

Structs

Delay
A future representing the notification that an elapsed duration has occurred.